refactor: separating jwt component
